fix: update Home Page on Read the Docs
Update Home Page on Read the Docs

Added links for all of the sports packages from API Documentation (sportsreference) and newly created FAQs page from the "/docs" subfolder onto to the Table of Contents in the Home Page. Created another Read the Docs link that I am doing for my version. This is an assignment for CS 464 - Open Source Software from Oregon State University in Fall 2022.

Signed-Off-By: Mallory Huston <hustonm@oregonstate.edu.>